+++ index.js
@@ -6,7 +6,10 @@ const {
 const { readDefinitionFile } = require("./src/lib/reader");
 const { parentChainFromNode } = require("./src/lib/util/chain-util");
 const { treatUrl } = require("./src/lib/util/reader-util");
-const { getBaseBranch } = require("./src/lib/util/mapping-util");
+const {
+  getBaseBranch,
+  getBaseMappingInfo
+} = require("./src/lib/util/mapping-util");
 
 module.exports = {
   getTree,
@@ -16,5 +19,6 @@ module.exports = {
   readDefinitionFile,
   parentChainFromNode,
   treatUrl,
-  getBaseBranch
+  getBaseBranch,
+  getBaseMappingInfo
 };