+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - [ISSUE 103] be able to access definition file from private repositories
 - [ISSUE 168] dependencies.exclude on project-dependencies tests added
 - getBaseBranch mechanism
+- getBaseMappingInfo mechanism
 
 # V2.1.0